+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/ChatClient.kt
@@ -61,6 +61,7 @@ import io.getstream.chat.android.client.errorhandler.onReactionError
 import io.getstream.chat.android.client.errors.ChatError
 import io.getstream.chat.android.client.events.ChatEvent
 import io.getstream.chat.android.client.events.ConnectedEvent
+import io.getstream.chat.android.client.events.ConnectingEvent
 import io.getstream.chat.android.client.events.DisconnectedEvent
 import io.getstream.chat.android.client.events.HasOwnUser
 import io.getstream.chat.android.client.events.NewMessageEvent
@@ -87,6 +88,7 @@ import io.getstream.chat.android.client.models.BannedUser
 import io.getstream.chat.android.client.models.BannedUsersSort
 import io.getstream.chat.android.client.models.Channel
 import io.getstream.chat.android.client.models.ConnectionData
+import io.getstream.chat.android.client.models.ConnectionState
 import io.getstream.chat.android.client.models.Device
 import io.getstream.chat.android.client.models.EventType
 import io.getstream.chat.android.client.models.Filters
@@ -124,6 +126,8 @@ import io.getstream.chat.android.client.plugin.listeners.ShuffleGiphyListener
 import io.getstream.chat.android.client.plugin.listeners.ThreadQueryListener
 import io.getstream.chat.android.client.plugin.listeners.TypingEventListener
 import io.getstream.chat.android.client.setup.InitializationCoordinator
+import io.getstream.chat.android.client.setup.state.ClientState
+import io.getstream.chat.android.client.setup.state.internal.toMutableState
 import io.getstream.chat.android.client.socket.ChatSocket
 import io.getstream.chat.android.client.socket.SocketListener
 import io.getstream.chat.android.client.token.CacheableTokenProvider
@@ -201,6 +205,12 @@ internal constructor(
         }
     )
 
+    /**
+     * With clientState allows the user to get the state of the SDK like connection, initialization...
+     */
+    public val clientState: ClientState
+        get() = ClientState.get()
+
     private var pushNotificationReceivedListener: PushNotificationReceivedListener =
         PushNotificationReceivedListener { _, _ -> }
 
@@ -230,6 +240,12 @@ internal constructor(
                     }
                     api.setConnection(user.id, connectionId)
                     notifications.onSetUser()
+
+                    clientState.toMutableState()?.run {
+                        setConnectionState(ConnectionState.CONNECTED)
+                        setInitialized(true)
+                        setUser(user)
+                    }
                 }
                 is DisconnectedEvent -> {
                     when (event.disconnectCause) {
@@ -242,11 +258,15 @@ internal constructor(
                             if (ToggleService.isSocketExperimental().not()) {
                                 socketStateService.onSocketUnrecoverableError()
                             }
+                            clientState.toMutableState()?.setConnectionState(ConnectionState.OFFLINE)
                         }
                     }
                 }
                 is NewMessageEvent -> {
                     notifications.onNewMessageEvent(event)
+                }
+                is ConnectingEvent -> {
+                    clientState.toMutableState()?.setConnectionState(ConnectionState.CONNECTING)
                 }
                 else -> Unit // Ignore other events
             }
@@ -335,6 +355,9 @@ internal constructor(
         val isAnonymous = user == anonUser
         val cacheableTokenProvider = CacheableTokenProvider(tokenProvider)
         val userState = userStateService.state
+
+        ClientState.get().toMutableState()?.setUser(user)
+
         return when {
             tokenUtils.getUserId(cacheableTokenProvider.loadToken()) != user.id -> {
                 logger.logE(
@@ -998,6 +1021,7 @@ internal constructor(
         logger.logI("[disconnect] no args")
         notifications.onLogout()
         // fire a handler here that the chatDomain and chatUI can use
+        clientState.toMutableState()?.clearState()
         getCurrentUser().let(initializationCoordinator::userDisconnected)
         if (ToggleService.isSocketExperimental().not()) {
             socketStateService.onDisconnectRequested()